Experimented with multiple exec_next calls per step (now inactive).
The simulation can be sped up a little bit by making multiple
"exec_next" calls in succession without simulating the rest of the
of the computer. However, the speed increase is not that significant,
and it could lead to incorrect interpretation of the state. So while
it was tried, it is currently commented out.

For reference, the implementation made multiple calls whenever the
speed setting (wait interval) was zero.