Show PC at current instruction (even though it has already advanced).
The previous version moved the PC display pointer to reflect the
program counter's current value. However, the PC is advanced early
in the series of microcode steps for each instruction. This can be
confusing when trying to follow the microcode and program flow.

This change introduces a "last_pc" value to keep track of the last
value for the PC and only change it on step 0 of the microcode
counter. This seems to work properly even during jump instructions.