Removed index.html (which was identical to Ben_Eater_8bit.html)
The easiest way to publish live pages on GitHub is to use a file
named "index.html" as the main page. So "Ben_Eater_8bit.html" was
copied to index.html to get it published. However, it would be
less error prone to have just one file, so "Ben_Eater_8bit.html"
will be renamed to "index.html" in a subsequent commit.